requirements.txt
torch
transformers
accelerate
safetensors
Pillow
opencv-python
git+https://github.com/huggingface/diffusers